refactor(core): share the attachment add cancellation machinery
