+++ .changeset/attachment-add-operations.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/core": patch
+---
+
+refactor: share the attachment add cancellation machinery between the composer core and the store client.
